Add files via upload
mtes

+++ project-spring-jsf/src/main/java/com/jsf/controller/BusController.java
@@ -1,8 +1,8 @@
 package com.jsf.controller;
 
-
 import java.io.Serializable;
 import java.util.ArrayList;
+import java.util.GregorianCalendar;
 import java.util.List;
 
 import javax.annotation.PostConstruct;
@@ -15,6 +15,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 
 import com.jsf.model.Bus;
 import com.jsf.service.BusService;
+import com.jsf.vo.BusVO;
 
 @Named
 @ViewScoped
@@ -25,54 +26,65 @@ public class BusController implements Serializable {
 	private static final long serialVersionUID = 581453877192605965L;
 
 	private Bus bus = new Bus();
-
+	private BusVO busVo = new BusVO();
 	private List<Bus> listBus = new ArrayList<>();
+	private List<BusVO> listBusVo = new ArrayList<>();
 
 	@Autowired
 	private BusService busService;
 
 	@PostConstruct
-    public void init(){
+	public void init() {
 		fetchAll();
-    }
-	
+		refresh();
+	}
+
 	public void fetchAll() {
-		listBus = busService.findAll();
+		listBusVo = busService.fetchAll();
+		refresh();
 	}
 
 	public void save() {
 		String message = "";
-		if(busService.validation()) {
-			busService.save(bus);
-			bus = new Bus();
+		if (busService.validation()) {
+			busService.saveBus(busVo);
 			fetchAll();
 			message = "Succes Save";
-		}else {
+		} else {
+			fetchAll();
 			message = "Failed Save";
 		}
 		message(message);
-			
+
 	}
-	
+
+	public void delete(BusVO busVo) {
+		bus = busService.findBusById(busVo);
+		if (bus != null) {
+			busService.delete(bus);
+			fetchAll();
+			message("Succes Delete");
+		} else {
+			message("Failed Delete");
+		}
+	}
+
+	public void edit(BusVO bus) {
+		this.busVo = bus;
+	}
+
+	public void refresh() {
+		busVo = new BusVO();
+		busVo.setScheduledDateTime(GregorianCalendar.getInstance().getTime());
+		busVo.setActualDateTime(GregorianCalendar.getInstance().getTime());
+		busVo.setEstimatedDateTime(GregorianCalendar.getInstance().getTime());
+	}
+
 	public void message(String message) {
 		FacesContext context = FacesContext.getCurrentInstance();
 		context.addMessage(null,
 				new FacesMessage(message.contains("Failed") ? FacesMessage.SEVERITY_ERROR : FacesMessage.SEVERITY_INFO,
 						"Message", message));
-	}
-
-	public void delete(Bus bus) {
-		busService.delete(bus);
-		fetchAll();
-		message("Succes Delete");
-	}
-	
-	public void edit(Bus bus) {
-		this.bus = bus;
-	}
-
-	public void refresh() {
-		bus = new Bus();
 	}
 
 	public Bus getBus() {
@@ -98,5 +110,21 @@ public class BusController implements Serializable {
 	public void setBusService(BusService busService) {
 		this.busService = busService;
 	}
-	
+
+	public BusVO getBusVo() {
+		return busVo;
+	}
+
+	public void setBusVo(BusVO busVo) {
+		this.busVo = busVo;
+	}
+
+	public List<BusVO> getListBusVo() {
+		return listBusVo;
+	}
+
+	public void setListBusVo(List<BusVO> listBusVo) {
+		this.listBusVo = listBusVo;
+	}
+
 }
